Use 32bit ids because sqlite doesn't support u64 (and it seems like 2^32 is enough)

+++ src/db.rs
@@ -12,16 +12,16 @@ pub trait DBFile {
     fn list_done(&self) -> Result<Vec<DbItem>>;
     fn list_undone(&self) -> Result<Vec<DbItem>>;
     fn insert(&self, item_name: &str) -> Result<()>;
-    fn delete(&self, id: u64) -> Result<()>;
-    fn get(&self, id: u64) -> Result<Option<DbItem>>;
+    fn delete(&self, id: u32) -> Result<()>;
+    fn get(&self, id: u32) -> Result<Option<DbItem>>;
     fn get_random(&self) -> Result<Option<DbItem>>;
-    fn done(&self, id: u64, time: NaiveDateTime) -> Result<()>;
-    fn undone(&self, id: u64) -> Result<()>;
+    fn done(&self, id: u32, time: NaiveDateTime) -> Result<()>;
+    fn undone(&self, id: u32) -> Result<()>;
     fn find(&self, item_name: &str) -> Result<Vec<DbItem>>;
 }
 
 pub struct DbItem {
-    pub id: u64,
+    pub id: u32,
     pub name: String,
     pub completed_at: Option<NaiveDateTime>,
 }
